+++ astro.config.mjs
@@ -1,4 +1,4 @@
-import { defineConfig, envField } from "astro/config";
+import { defineConfig } from "astro/config";
 import tailwind from "@astrojs/tailwind";
 import mdx from "@astrojs/mdx";
 import sitemap from "@astrojs/sitemap";
@@ -8,42 +8,7 @@ import dotenv from "dotenv";
 dotenv.config();
 
 export default defineConfig({
-  site: process.env.BASE_URL ?? "https://himanshubhardwaz.com",
-  vite: {
-    build: {
-      minify: false,
-    },
-  },
-  env: {
-    schema: {
-      TURSO_DATABASE_URL: envField.string({
-        required: true,
-        context: "server",
-        access: "secret",
-      }),
-      TURSO_AUTH_TOKEN: envField.string({
-        required: true,
-        context: "server",
-        access: "secret",
-      }),
-      BASE_URL: envField.string({
-        required: true,
-        context: "client",
-        access: "public",
-        default: "https://himanshubhardwaz.com",
-      }),
-      GOOGLE_OAUTH_CLIENT_ID: envField.string({
-        required: true,
-        context: "server",
-        access: "secret",
-      }),
-      GOOGLE_OAUTH_CLIENT_SECRET: envField.string({
-        required: true,
-        context: "server",
-        access: "secret",
-      }),
-    },
-  },
+  site: process.env.SITE_URL,
   adapter: cloudflare(),
   integrations: [tailwind(), mdx(), sitemap()],
   markdown: {
@@ -54,7 +19,7 @@ export default defineConfig({
       },
     },
   },
-  output: "server",
+  output: "hybrid",
   vite: {
     define: {
       "process.env": process.env,
