+++ drizzle.config.ts
@@ -1,14 +1,12 @@
-import { config } from "dotenv";
-import { defineConfig } from "drizzle-kit";
+import type { Config } from "drizzle-kit";
 
-config({ path: ".env" });
-
-export default defineConfig({
-  schema: "./src/lib/server/schema.ts",
+export default {
+  schema: "./src/db/schema.ts",
   out: "./migrations",
-  dialect: "turso",
+  dialect: "sqlite",
+  driver: "turso",
   dbCredentials: {
     url: process.env.TURSO_DATABASE_URL!,
-    authToken: process.env.TURSO_AUTH_TOKEN!,
+    authToken: process.env.TURSO_AUTH_TOKEN,
   },
-});
+} satisfies Config;
