+++ package.json
@@ -13,32 +13,27 @@
     "db:studio": "drizzle-kit studio"
   },
   "dependencies": {
-    "@astrojs/mdx": "^4.0.3",
-    "@astrojs/rss": "^4.0.10",
+    "@astrojs/mdx": "^3.1.9",
+    "@astrojs/rss": "^4.0.9",
     "@astrojs/sitemap": "^3.2.1",
-    "@astrojs/svelte": "^7.0.2",
-    "@astrojs/tailwind": "^5.1.4",
+    "@astrojs/svelte": "^5.7.2",
+    "@astrojs/tailwind": "^5.1.2",
     "@fontsource-variable/inter": "^5.1.0",
     "@fontsource-variable/jetbrains-mono": "^5.1.1",
     "@libsql/client": "^0.10.0",
-    "@oslojs/crypto": "^1.0.1",
-    "@oslojs/encoding": "^1.1.0",
-    "arctic": "^2.3.2",
-    "astro": "^5.1.1",
+    "astro": "^4.16.8",
     "astro-seo": "^0.8.4",
     "dotenv": "^16.4.7",
-    "drizzle-orm": "^0.38.2",
-    "nanostores": "^0.11.3",
+    "drizzle-orm": "^0.33.0",
     "sharp": "^0.33.5",
     "tailwindcss": "^3.4.14"
   },
   "devDependencies": {
     "@astrojs/check": "^0.9.4",
-    "@astrojs/cloudflare": "^12.1.0",
+    "@astrojs/cloudflare": "^11.2.0",
     "@flydotio/dockerfile": "^0.5.9",
     "@tailwindcss/typography": "^0.5.15",
-    "drizzle-kit": "^0.30.1",
+    "drizzle-kit": "^0.24.2",
     "typescript": "^5.6.3"
-  },
-  "packageManager": "pnpm@8.15.4+sha1.c85a4305534f76d461407b59277b954bac97b5c4"
+  }
 }
