--- src/components/LoginOrLogoutButton.astro
@@ -1,45 +1,0 @@
----
-import Button from "./Button.astro";
-
-const token = Astro?.cookies?.get("session")?.value ?? null;
----
-
-<Button id="logout" style={`display: ${token ? "block" : "none"};`}>
-  Logout
-</Button>
-<Button
-  href="/login"
-  id="login"
-  style={`display: ${token ? "none" : "block"};`}
->
-  Login
-</Button>
-
-<script>
-  const setupLogoutListener = () => {
-    const logout = document.getElementById("logout");
-
-    if (logout) {
-      logout.addEventListener("click", async () => {
-        try {
-          const response = await fetch("/logout", { method: "POST" });
-          if (response.ok) {
-            const logoutButton = document.getElementById("logout");
-            const loginButton = document.getElementById("login");
-            if (loginButton) loginButton.style.display = "block";
-            if (logoutButton) logoutButton.style.display = "none";
-          }
-        } catch (error) {
-          console.error("Logout error:", error);
-        }
-      });
-      return true;
-    }
-    return false;
-  };
-
-  const interval = setInterval(() => {
-    const setup = setupLogoutListener();
-    if (setup) clearInterval(interval);
-  }, 50);
-</script>
