+++ src/components/TechPill.astro
@@ -6,9 +6,6 @@ interface Props {
 
 const { id } = Astro.props;
 const tech = await getEntry("techs", id);
-if (!tech) {
-  return;
-}
 const { name, href } = tech.data;
 const { Content } = await tech.render();
 ---
