--- src/lib/server/oauth.ts
@@ -1,11 +1,0 @@
-import { Google } from "arctic";
-import {
-  GOOGLE_OAUTH_CLIENT_ID,
-  GOOGLE_OAUTH_CLIENT_SECRET,
-} from "astro:env/server";
-
-export const google = new Google(
-  GOOGLE_OAUTH_CLIENT_ID,
-  GOOGLE_OAUTH_CLIENT_SECRET,
-  "http://localhost:4321/login/google/callback"
-);
