--- src/lib/server/session.ts
@@ -1,91 +1,0 @@
-import type { User, Session } from "~/lib/server/db";
-import { db, sessionTable, userTable } from "~/lib/server/db";
-import {
-  encodeBase32LowerCaseNoPadding,
-  encodeHexLowerCase,
-} from "@oslojs/encoding";
-import { sha256 } from "@oslojs/crypto/sha2";
-import { eq } from "drizzle-orm/expressions";
-import type { APIContext } from "astro";
-
-export type SessionValidationResult =
-  | { session: Session; user: User }
-  | { session: null; user: null };
-
-export function generateSessionToken(): string {
-  const bytes = new Uint8Array(20);
-  crypto.getRandomValues(bytes);
-  const token = encodeBase32LowerCaseNoPadding(bytes);
-  return token;
-}
-
-export async function createSession(
-  token: string,
-  userId: number
-): Promise<Session> {
-  const sessionId = encodeHexLowerCase(sha256(new TextEncoder().encode(token)));
-  const session: Session = {
-    id: sessionId,
-    userId,
-    expiresAt: new Date(Date.now() + 1000 * 60 * 60 * 24 * 30),
-  };
-  await db.insert(sessionTable).values(session);
-  return session;
-}
-
-export async function validateSessionToken(
-  token: string
-): Promise<SessionValidationResult> {
-  const sessionId = encodeHexLowerCase(sha256(new TextEncoder().encode(token)));
-  const result = await db
-    .select({ user: userTable, session: sessionTable })
-    .from(sessionTable)
-    .innerJoin(userTable, eq(sessionTable.userId, userTable.id))
-    .where(eq(sessionTable.id, sessionId));
-  if (result.length < 1) {
-    return { session: null, user: null };
-  }
-  const { user, session } = result[0];
-  if (Date.now() >= session.expiresAt.getTime()) {
-    await db.delete(sessionTable).where(eq(sessionTable.id, session.id));
-    return { session: null, user: null };
-  }
-  if (Date.now() >= session.expiresAt.getTime() - 1000 * 60 * 60 * 24 * 15) {
-    session.expiresAt = new Date(Date.now() + 1000 * 60 * 60 * 24 * 30);
-    await db
-      .update(sessionTable)
-      .set({
-        expiresAt: session.expiresAt,
-      })
-      .where(eq(sessionTable.id, session.id));
-  }
-  return { session, user };
-}
-
-export async function invalidateSession(sessionId: string): Promise<void> {
-  await db.delete(sessionTable).where(eq(sessionTable.id, sessionId));
-}
-
-export function setSessionTokenCookie(
-  context: APIContext,
-  token: string,
-  expiresAt: Date
-): void {
-  context.cookies.set("session", token, {
-    httpOnly: true,
-    sameSite: "lax",
-    secure: import.meta.env.PROD,
-    expires: expiresAt,
-    path: "/",
-  });
-}
-
-export function deleteSessionTokenCookie(context: APIContext): void {
-  context.cookies.set("session", "", {
-    httpOnly: true,
-    sameSite: "lax",
-    secure: import.meta.env.PROD,
-    maxAge: 0,
-    path: "/",
-  });
-}
