--- src/pages/_components/LikesViewsCounter.astro
@@ -1,21 +1,0 @@
----
-const { slug } = Astro.props;
-
-export const prerender = false;
-
-export let likeCount = 0;
-export let viewCount = 0;
-
-import { BASE_URL } from "astro:env/client";
-
-try {
-  const response = await fetch(`${BASE_URL}/api/blog/${slug}`);
-  const { likes, views } = await response.json();
-  likeCount = likes;
-  viewCount = views;
-} catch (error) {
-  console.error(error);
-}
----
-
-<h5>Likes: {likeCount}; Views: {viewCount}</h5>
