+++ src/pages/blog/[slug].astro
@@ -4,11 +4,10 @@ import { Image } from "astro:assets";
 import { getCollection } from "astro:content";
 import Button from "~/components/Button.astro";
 import Prose from "~/components/Prose.astro";
-import Layout from "~/layouts/Layout.astro";
+import BaseLayout from "~/layouts/Layout.astro";
 import { components } from "~/components/mdx";
 import "@fontsource-variable/jetbrains-mono";
 import jetbrainsMonoFile from "@fontsource-variable/jetbrains-mono/files/jetbrains-mono-latin-wght-normal.woff2?url";
-import LikesViewsCounter from "~/pages/_components/LikesViewsCounter.astro";
 
 export const getStaticPaths = (async () => {
   return (await getCollection("posts")).map((article) => ({
@@ -20,9 +19,21 @@ export const getStaticPaths = (async () => {
 const { article } = Astro.props;
 const { data, slug } = article;
 const { Content } = await article.render();
+
+export let likeCount = 0;
+export let viewCount = 0;
+
+try {
+  const response = await fetch(`${import.meta.env.SITE}/api/blog/${slug}`);
+  const { likes, views } = await response.json();
+  likeCount = likes;
+  viewCount = views;
+} catch (error) {
+  console.error(error);
+}
 ---
 
-<Layout
+<BaseLayout
   title={data.title}
   description={data.description}
   image={data.cover.src.src}
@@ -77,11 +88,7 @@ const { Content } = await article.render();
         transition:name={`article-cover:${article.slug}`}
       />
       <h1 transition:name={`article-title:${article.slug}`}>{data.title}</h1>
-
-      <LikesViewsCounter slug={slug} server:defer>
-        <p slot="fallback">Loading...</p>
-      </LikesViewsCounter>
-
+      <h5>Likes: {likeCount}; Views: {viewCount}</h5>
       <p class="lead" transition:name={`article-description:${article.slug}`}>
         {data.description}
       </p>
@@ -124,10 +131,9 @@ const { Content } = await article.render();
       ></path>
     </svg>
   </button>
-</Layout>
+</BaseLayout>
 
 <script>
-  import { BASE_URL } from "astro:env/client";
   import { getSlugFromURL } from "~/utils/getSlugFromURL";
   import type { Article } from "~/types";
 
@@ -213,13 +219,16 @@ const { Content } = await article.render();
 
   if (!isArticleViewed()) {
     try {
-      const response = await fetch(`${BASE_URL}/api/blog/${getSlugFromURL()}`, {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify({ viewed: true }),
-      });
+      const response = await fetch(
+        `${import.meta.env.SITE}/api/blog/${getSlugFromURL()}`,
+        {
+          method: "POST",
+          headers: {
+            "Content-Type": "application/json",
+          },
+          body: JSON.stringify({ viewed: true }),
+        }
+      );
       const { viewed } = await response.json();
       updatedViewState(viewed);
     } catch (error) {
@@ -244,13 +253,16 @@ const { Content } = await article.render();
     if (likedSvg) likedSvg.classList.toggle("liked", liked);
 
     try {
-      const response = await fetch(`${BASE_URL}/api/blog/${getSlugFromURL()}`, {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify({ liked }),
-      });
+      const response = await fetch(
+        `${import.meta.env.SITE}/api/blog/${getSlugFromURL()}`,
+        {
+          method: "POST",
+          headers: {
+            "Content-Type": "application/json",
+          },
+          body: JSON.stringify({ liked }),
+        }
+      );
       const { liked: updatedLiked } = await response.json();
       updateLikeState(updatedLiked);
       if (likedSvg) likedSvg.classList.toggle("liked", updatedLiked);
