+++ src/pages/index.astro
@@ -1,6 +1,4 @@
 ---
-export const prerender = false;
-
 import "~/styles/main.css";
 import Hero from "./_components/Hero.astro";
 import About from "./_components/About.astro";
@@ -8,10 +6,10 @@ import Experiences from "./_components/Experiences.astro";
 import MoreProjects from "./_components/MoreProjects.astro";
 import LatestArticle from "./_components/LatestArticle.astro";
 import Contact from "./_components/Contact.astro";
-import Layout from "~/layouts/Layout.astro";
+import BaseLayout from "~/layouts/Layout.astro";
 ---
 
-<Layout
+<BaseLayout
   title="Home"
   description="Software Engineer, Full Stack Developer, Frontend Developer"
 >
@@ -21,4 +19,4 @@ import Layout from "~/layouts/Layout.astro";
   <MoreProjects />
   <LatestArticle />
   <Contact />
-</Layout>
+</BaseLayout>
